+++ CHANGES.md
@@ -42,6 +42,7 @@ Changes in 12.5
 - JS remove inline code from program in ChooseCourse.php & ChooseCourse.js
 - JS remove inline code from program in ChooseRequest.php & ChooseRequest.js
 - JS remove inline code from program in Export.php & Export.js
+- JS remove inline code from program in Courses.php & Courses.js
 
 Changes in 12.4.3
 -----------------